requirements.txt
xknx
python-can